Refactored names, created new folder with components
