+++ src/DesktopComponents/ExploreMarketplaceComponent/ExploreMarketplaceFolder/CardCase/CardCase.tsx
@@ -1,11 +1,11 @@
-import classes from "../exploreMarketplace.module.scss";
-import Card from "../card/card.tsx";
-import {CardSlice} from "../../../reduxTest/cards/CardSlice.ts";
+import classes from "../ExploreMarketplace.module.scss";
+import Card from "../Card/card.tsx";
+import {CardSlice} from "../../../../reduxTest/cards/CardSlice.ts";
 import {connect, ConnectedProps} from "react-redux";
-import {fetchCardsAsync} from "../../../client/asyncThunk/fetchCardsAsync.ts";
+import {fetchCardsAsync} from "../../../../Api/AsyncThunk/FetchCardsAsync.ts";
 /*import {useState} from "react";*/
-/*import Pagination from "../paginator/Paginator.tsx";*/
-import {RootState} from "../../../reduxTest/store/Store.ts";
+/*import Pagination from "../Paginator/Paginator.tsx";*/
+import {RootState} from "../../../../reduxTest/store/Store.ts";
 
 //Закомментировал Пагинатор
 //В {CardArray.map((cardItem: CardSlice) => ( заменить CardArray на currentItems чтобы работало
