+++ src/DesktopComponents/FooterComponent/Footer.module.scss
