+++ src/DesktopComponents/TopBannerComponent/TopBanner.module.scss
@@ -49,7 +49,7 @@ $max-width: 375px;
     left: 0;
     width: 100%;
     height: 84%;
-    background: url(".././IMG/Rectangle3.png") no-repeat center/cover;
+    background: url("../../IMG/Rectangle3.png") no-repeat center/cover;
     filter: blur(45px);
     z-index: 0;
   }
@@ -104,7 +104,7 @@ position: relative;
     left: 0;
     width: 100%;
     height: 8%;
-    background: url(".././IMG/Rectangle3Small.png") no-repeat center/cover;
+    background: url("../../IMG/Rectangle3Small.png") no-repeat center/cover;
     filter: blur(36px);
     z-index: 0;
   }
