+++ src/DesktopComponents/TopCollectionComponent/TopCollectionComponent.tsx
@@ -1,0 +1,10 @@
+import TopCollection from "./TopCollectionFolder/TopCollection.tsx";
+import classes from "../../App.module.scss";
+
+const TopCollectionComponent = () => {
+    return (<div className={classes.contentBack}>
+            <TopCollection/>
+        </div>
+    )
+}
+export default TopCollectionComponent;
