+++ src/MobileComponents/MobileCardsPreview/CardsPreview.tsx
@@ -1,4 +1,4 @@
-import {HeroPictures} from "../../Hero/hero.tsx";
+import {HeroPictures} from "../../DesktopComponents/TopBannerComponent/TopBannerComponent.tsx";
 
 const CardsPreview = () => {
     return (<div style={{marginLeft: 30, marginBottom: 110}}>
