--- src/content2/Content2.tsx
@@ -1,10 +1,0 @@
-import TopCollection from "./topCollection/TopCollection.tsx";
-import classes from "../App.module.scss";
-
-const Content2 = () => {
-    return (<div className={classes.contentBack}>
-            <TopCollection/>
-        </div>
-    )
-}
-export default Content2
