+++ src/pages/creators/pageCreators.tsx
@@ -1,12 +1,12 @@
-import Menu from "../../menu/Menu.tsx";
-import Footer from "../../footer/footer.tsx";
+import NavigationComponent from "../../DesktopComponents/NavigationComponent/NavigationComponent.tsx";
+import FooterComponent from "../../DesktopComponents/FooterComponent/FooterComponent.tsx";
 
 
 const PageCreators = () =>{
     return (
         <div>
-            <Menu/>
-            <Footer/>
+            <NavigationComponent/>
+            <FooterComponent/>
         </div>
     )
 }
